Add 'try-load' sub-command
Introduce the `try-load` sub-command that could also be aliased
`try-add`. try-load attempts to load modulefile(s) but does not raise
error (message or code) if module(s) cannot be found (does not exist or
forbidden).

Error is still raised if an evaluation error occurs.

The `try_modulefile` internal state is added to transmit the information
that the load (or unload when load is proccessed during an unload
evaluation) is performed in *try* mode or not.