Add conflict_unload configuration option
Add conflict_unload configuration option that controls whether or not
conflicting modules and their dependent should be removed when targeted
by "module unload" or "conflict" modulefile commands or when loading a
module which is defined as a conflict by already loaded modules.

"auto_handling" should also be enabled to unload conflicts on the two
later cases.