Fix detection of already loaded module with variant set
When an attempt is made to load a module that is already loaded with
some variants defined, this loaded module were not correctly detected
and a "another variant is loaded" error was obtained instead of "module
is already loaded" notification.

Parsing of loaded module specification (with variant information) was
missing thus subsequent loaded module checks were not performed correctly.