Drop failed module eval when dropping held reports
When report hold is enabled (during evaluation of satisfying
requirement), also track failed module evaluation. If held report is
dropped, also drop the failed module evaluation that occurred during the
held time.

With this change, failed evaluation that happened during a Requirement
Load mechanism but not reported as an alternative is successfully
loaded. If this module is then later tried again, the evaluation will
not be skipped and this later context will get the accurate error
information.