Separate processing of DepRe mods of UReqUn
This commit is part of a global change to the way the Useless
Requirement Unload modules are computed and proceeded.

Split processing of Dependent Reload modules to treat DepRe modules
coming from UReqUn modules separately, right before unloading the UReqUn
modules.

As a result, unload order may vary, but it only concern the DepRe
modules coming from UReqUn modules, which means modules in conflict with
UReqUn modules that benefits from a reload due to the unload of the
UReqUns.

One limitation to note: UReqUn modules still need to be computed at the
beginning of the auto_handling processing block, to withdraw them from
the list of the initial DepRe module list.