Fix getRequiredLoadedModuleList and add argument
Fix parsing of dependency structure in getRequiredLoadedModuleList where
entries in g_moduleDepend array are a list of lists. Each containing all
loaded alternatives for a given requirement.

Add excluded_mod_list optional argument to getRequiredLoadedModuleList
procedure to precise a list of modules to exclude (usually to skip
requirements that will be unloaded).