Add UReqUn process on top Load evaluation
In case of a top Load evaluation (load or switch command), apply the
Useless Requirement Unload mechanism after the load of the top module
(and prior the load phase of the Dependent Reload modules).

Like for Unload evaluation, a list of DepRe modules for these UReqUn
modules is built and added to the global DepRe modules list.