Apply UReqUn process on Unload only for top eval
Apply the Useless Requirement Unload process on Unload only if this is a
top evaluation and not a switch (in which case the load phase handles
the UReqUn process).

With this change, the UReqUn process is corrected as it sometimes
unloaded modules that were later on Requirement for loading modules
(especially when treating unload of Conflict Unload modules).