Optimize use of return in procedure
Optimize procedure to immediately return rather set a common variable
that is returned at the end of the procedure. This change is applied
onto every procedure that do not progressively build a result to return.

Also simplify procedures that returns an empty string in case a variable
is not defined. Use the implicit return value of a procedure, which is
the value returned by last command executed in the proc. Conditional
commands like if or switch return an empty string if no default block
(else or default) is defined for them.

To illustrate last point, consider the following code:

proc myproc {} {
   if {[info exists myvar]} {
      return $myvar
   }
}

If myvar is not defined, myproc will return the value returned by the if
command. As there is no else block for this if command, if returns an
empty string. Thus proc returns an empty string if myvar is not defined.