No DepRe list update if unload phase fail
When unloading Dependent Reload modules, if one unload fails an error is
raised. If force mode is enabled, no evaluation may fail (as
super-sticky modules are allowed to be reloaded). So there is no need to
update DepRe module list as a failure will lead to a global error that
will stop processing.