Apply DepRe load phase only on top load/unload/switch
Proceed with Dependent Reload load phase on load, unload or switch
evaluation only when these evaluations are top level ones.

As a consequence DepRe modules are reloaded after top module load,
unload or switch. They were previously reloaded right after the load,
unload or switch of the module that triggered the DepRe unload phase.

In case a DepRe is a ReqLo for another module to load, it will be loaded
as ReqLo. Currently it looses its extra tags and gets the auto-loaded
tag. This is something to fix in the future.

With this change when a DepRe module triggered by a sub module switch
operation fails to reload, a real error is rendered rather a simple
warning.

As all DepRe are reloaded at the end, some of them that could be
qualified as UReqUn skip this step and are not auto unloaded.