Allow extra match search on "module use" command
Add the ability to perform an extra match search on the "use"
sub-command of module command in modulefile.

Also consider direct edition of the MODULEPATH environment variable
with append-path or prepend-path modulefile commands as result for "use"
extra match element.

Modulepath arguments set on these commands are recorded in their
absolute path form.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>