Introduce "spider" sub-command
Add the "spider" sub-command that scans all the enabled modulepaths and
the modulepaths enabled by modules in every modulepaths to return the
modules matching the search query.

All kind of options and search queries supported by avail sub-command
are supported by spider.

spider is implemented as a two-step process:
* find all modulepaths through a scan evaluation to fetch all "use"
  elements
* perform an avail on all these modulepaths

Properties of the spider sub-command:
* insensitive match search enabled
* shortest shortcut: "spi"
* --json output support
* --all/-a to report hidden modules
* always_read_full_file: true
* allow_noname_spec: true
* allow_xt_spec: true
* parse arguments as module specification: true
* only callable from top level: true
* accept_minus_arg: true
* collect tags: true

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>