Pre-enabled modulepath not considered reachable via loaded mod
In case a modulepath is already set in the MODULEPATH list before a
loaded module enables it again, this modulepath is not considered
reachable via the loaded module.

Reference counter set, when additionally using the modulepath by
modulefiles, is used to determine if modulepath was set prior loaded
module(s) enabling it.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>