Change approach with version req in cache file magic cookie
Version in magic cookie of cache files was previously set to the current
version of Modules. As a result, only the same Modules version would be
able to use cache file generated.

This is changed to enable use of cache files by several Modules release.
Only a cache format change would imply a magic cookie version upgrade.

So as of now, the version required is set to 5.3. This will change if
the cache format evolves (with new cache file Tcl commands or a change
in the arguments of the existing commands).

Version requirement is set to 5.3 as commands in cache files as not been
changed since the introduction of the cache feature in Modules 5.3
release.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>