Attempt mod reload if modpath is enabled or mod still avail
When reloading Dependent Reload modules, do not attempt to reload
modules whose modulepath is not enabled anymore and module not found in
enabled modulepaths.

Module availability is based on the search of module name and version
file in the currently enabled modulepaths. No evaluation is made to
check if enabled variants are available in found modulefile.

Like for a Dependent Reload module with unsatisfied dependency, module
not available anymore in remaining enabled modulepaths becomes a
Dependent Unload module.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>