Insert provided-aliases in module search result
If provided-alias are requested, include them in the module search
result. Aliases defined with "provide" and "family" modulefile commands
are considered provided-aliases.

This change requires to perform the modulefile scan evaluation during
step #1 of getModules (gathering of entries) instead of step #4.
Provided-aliases are then inserted right after to be part of step #2 and
step #3 checks.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>