Add path_entry_reorder configuration option
Add path_entry_reorder configuration option that controls whether or not
a path-like environment variable should be updated if a path entry is
already present in the variable value.

If enabled, such entry is moved at the front or the back of the variable
value (depending on the modulefile command used: append-path or
prepend-path). This option when set matches the default behavior found
on Lmod.

Add path_entry_reorder configuration option to completion script.

Add the --enable-path-entry-reorder configure option to choose at
installation time the path_entry_reorder behavior. Default is 'n'.

When configuration is modified it sets the MODULES_PATH_ENTRY_REORDER
environment variable.

Fixes #595

Signed-off-by: Achim Gsell <achim.gsell@psi.ch>