Get correct tags for mod having same name than a loaded mod
Ensure that a module having the same name than a loaded module from
another modulepath will not inherit the tags of this loaded module.

Split collectModuleTags procedure to have a getMatchingTagList procedure
that gets the tags applying to a module. Use this procedure in
getTagList instead of the g_tagHash structure to avoid the tags only
applying to the loaded module if we are not checking this modulefile.

Fixes #599

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>