Drop in-memory tags when unloading module
When unloading a module, drop the tags attached to it which are stored
in memory. If this module is reloaded from another modulepath with other
tags set on it, the correct tag list will then be applied (and not
inherited from the tags applied to the unloaded version of the module).

This change requires to add g_tagHash structure to the list of arrays
that are saved before any evaluation and restored if a failure occurs.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>