Call cacheCurrentModules from single location
Rather making a call to cacheCurrentModules in every procedure that
needs to look at loaded environment, make a call to this procedure from
ml or moduleTop (unless chosen sub-command never read the loaded
environment).

cacheCurrentModules is simplified to always exit on error if an
inconsistency is detected in loaded environment. The lm_info_cached
state is removed as it is now useless.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>